+++ src/index.ts
@@ -2,7 +2,7 @@
 // Exports
 //
 
-export * from "./classes/FritterBodyParserBody.js";
 export * from "./classes/FritterBodyParserMiddleware.js";
+export * from "./classes/FritterBodyParserMiddlewareBody.js";
 
 export * from "./types/PossibleJSONData.js";